refactored form and entity form component
