+++ proxy.conf.json
@@ -1,6 +1,6 @@
 {
   "/db": {
-    "target": "https://dev.aam-digital.net/db",
+    "target": "http://localhost:3000",
     "secure": true,
     "logLevel": "debug",
     "changeOrigin": true,
