+++ src/app/core/entity-components/entity-details/form/form.component.html
@@ -1,0 +1,40 @@
+<div class="form-buttons-wrapper padding-bottom-small flex-row gap-small align-self-end">
+  <button
+    *ngIf="form?.enabled"
+    mat-raised-button
+    (click)="saveClicked()"
+    i18n="Save button for forms"
+  >
+    Save
+  </button>
+
+  <button
+    *ngIf="form?.enabled"
+    mat-raised-button
+    (click)="cancelClicked()"
+    i18n="Cancel button for forms"
+  >
+    Cancel
+  </button>
+
+  <ng-container *ngIf="form?.disabled">
+    <button
+      mat-raised-button
+      (click)="form.enable()"
+      *appDisabledEntityOperation="{
+        entity: entity,
+        operation: 'update'
+      }"
+      i18n="Edit button for forms"
+    >
+      Edit
+    </button>
+  </ng-container>
+</div>
+
+<app-entity-form
+  [entity]="entity"
+  [columns]="columns"
+  [columnHeaders]="headers"
+  (formReady)="initForm($event)"
+></app-entity-form>
