+++ src/app/core/entity-components/entity-details/form/form.component.scss
@@ -1,0 +1,15 @@
+@use "src/styles/variables/sizes";
+
+// Usually, this is the `:host` pseudo-selector but since the
+// view encapsulation is None, this has to be the name
+:host {
+  display: flex;
+  flex-direction: column;
+}
+
+.form-buttons-wrapper {
+  position: sticky;
+  top: sizes.$regular;
+  z-index: 100;
+  margin-right: 3px;
+}
