+++ src/app/core/entity-components/entity-details/form/form.component.ts
@@ -1,4 +1,4 @@
-import { Component } from "@angular/core";
+import { AfterViewInit, Component, ViewChild } from "@angular/core";
 import { OnInitDynamicComponent } from "../../../view/dynamic-components/on-init-dynamic-component.interface";
 import { PanelConfig } from "../EntityDetailsConfig";
 import { Entity } from "../../../entity/model/entity";
@@ -7,6 +7,16 @@ import { getParentUrl } from "../../../../utils/utils";
 import { Router } from "@angular/router";
 import { Location } from "@angular/common";
 import { DynamicComponent } from "../../../view/dynamic-components/dynamic-component.decorator";
+import { EntityFormComponent } from "../../entity-form/entity-form/entity-form.component";
+import { InvalidFormFieldError } from "../../entity-form/invalid-form-field.error";
+import {
+  EntityForm,
+  EntityFormService,
+} from "../../entity-form/entity-form.service";
+import { AlertService } from "../../../alerts/alert.service";
+import { filter } from "rxjs/operators";
+import { EntityMapperService } from "../../../entity/entity-mapper.service";
+import { ConfirmationDialogService } from "../../../confirmation-dialog/confirmation-dialog.service";
 
 /**
  * A simple wrapper function of the EntityFormComponent which can be used as a dynamic component
@@ -15,22 +25,26 @@ import { DynamicComponent } from "../../../view/dynamic-components/dynamic-compo
 @DynamicComponent("Form")
 @Component({
   selector: "app-form",
-  template: ` <app-entity-form
-    [entity]="entity"
-    [columns]="columns"
-    [columnHeaders]="headers"
-    [editing]="creatingNew"
-    (save)="saveClicked($event)"
-    (cancel)="cancelClicked()"
-  ></app-entity-form>`,
+  templateUrl: "./form.component.html",
+  styleUrls: ["./form.component.scss"],
 })
 export class FormComponent implements OnInitDynamicComponent {
   entity: Entity;
   columns: FormFieldConfig[][] = [];
   headers?: string[] = [];
   creatingNew = false;
+  saveInProgress = false;
+  form: EntityForm<Entity>;
+  private initialFormValues: any;
 
-  constructor(private router: Router, private location: Location) {}
+  constructor(
+    private router: Router,
+    private location: Location,
+    private entityFormService: EntityFormService,
+    private alertService: AlertService,
+    private entityMapper: EntityMapperService,
+    private confirmationDialog: ConfirmationDialogService
+  ) {}
 
   onInitFromDynamicConfig(config: PanelConfig) {
     this.entity = config.entity;
@@ -41,15 +55,71 @@ export class FormComponent implements OnInitDynamicComponent {
     }
   }
 
-  saveClicked(entity: Entity) {
-    if (this.creatingNew) {
-      this.router.navigate([getParentUrl(this.router), entity.getId()]);
+  async initForm(form: EntityForm<Entity>) {
+    await new Promise((resolve) => setTimeout(resolve));
+    this.form = form;
+    this.initialFormValues = this.form.getRawValue();
+    if (!this.creatingNew) {
+      this.form.disable();
     }
+    this.entityMapper
+      .receiveUpdates(this.entity.getConstructor())
+      .pipe(filter(({ entity }) => entity.getId() === this.entity.getId()))
+      .subscribe(({ entity }) => this.applyChanges(entity));
+  }
+
+  private async applyChanges(entity) {
+    if (this.saveInProgress || this.formIsUpToDate(entity)) {
+      // this is the component that currently saves the values -> no need to apply changes.
+      return;
+    }
+    if (
+      this.form.pristine ||
+      (await this.confirmationDialog.getConfirmation(
+        $localize`Load changes?`,
+        $localize`Local changes are in conflict with updated values synced from the server. Do you want the local changes to be overwritten with the latest values?`
+      ))
+    ) {
+      this.resetForm(entity);
+    }
+  }
+
+  private formIsUpToDate(entity: Entity): boolean {
+    return Object.entries(this.form.getRawValue()).every(
+      ([key, value]) =>
+        entity[key] === value || (entity[key] === undefined && value === null)
+    );
+  }
+
+  async saveClicked() {
+    this.saveInProgress = true;
+    try {
+      await this.entityFormService.saveChanges(this.form, this.entity);
+      this.form.markAsPristine();
+      this.form.disable();
+      if (this.creatingNew) {
+        this.router.navigate([getParentUrl(this.router), this.entity.getId()]);
+      }
+    } catch (err) {
+      if (!(err instanceof InvalidFormFieldError)) {
+        this.alertService.addDanger(err.message);
+      }
+    }
+    // Reset state after a short delay
+    setTimeout(() => (this.saveInProgress = false), 1000);
   }
 
   cancelClicked() {
     if (this.creatingNew) {
       this.location.back();
     }
+    this.resetForm();
+    this.form.disable();
+  }
+
+  private resetForm(entity = this.entity) {
+    // Patch form with values from the entity
+    this.form.patchValue(Object.assign(this.initialFormValues, entity));
+    this.form.markAsPristine();
   }
 }
