+++ src/app/core/entity-components/entity-form/entity-form/entity-form.component.html
@@ -1,43 +1,3 @@
-<!-- The wrapper for edit, save and cancel button.
-
-The classes 'flex-row' and 'gap-small' enable a small gap between the Save button and the Cancel button.
-Simply applying a margin right to the 'Save' button collides with the styles that the 'mat-raised-button'
-brings with it.
--->
-<div class="form-buttons-wrapper padding-bottom-small flex-row gap-small align-self-end">
-  <button
-    *ngIf="form?.enabled"
-    mat-raised-button
-    (click)="saveForm()"
-    i18n="Save button for forms"
-  >
-    Save
-  </button>
-
-  <button
-    *ngIf="form?.enabled"
-    mat-raised-button
-    (click)="cancelClicked()"
-    i18n="Cancel button for forms"
-  >
-    Cancel
-  </button>
-
-  <ng-container *ngIf="form?.disabled">
-    <button
-      mat-raised-button
-      (click)="this.form.enable()"
-      *appDisabledEntityOperation="{
-        entity: entity,
-        operation: 'update'
-      }"
-      i18n="Edit button for forms"
-    >
-      Edit
-    </button>
-  </ng-container>
-</div>
-
 <form class="entity-form-columns">
   <div *ngFor="let col of _columns; let i=index" class="entity-form-cell">
     <h2 *ngIf="columnHeaders?.[i]">{{columnHeaders[i]}}</h2>
