+++ src/app/core/entity-components/entity-form/entity-form/entity-form.component.scss
@@ -1,12 +1,4 @@
 @use "src/styles/mixins/grid-layout";
-@use "src/styles/variables/sizes";
-
-// Usually, this is the `:host` pseudo-selector but since the
-// view encapsulation is None, this has to be the name
-app-entity-form {
-  display: flex;
-  flex-direction: column;
-}
 
 .entity-form-columns {
   @include grid-layout.adaptive($min-block-width: 330px, $max-screen-width: 414px);
@@ -37,11 +29,4 @@ app-entity-form {
 
 .tooltip-suffix {
   margin: 4px;
-}
-
-.form-buttons-wrapper {
-  position: sticky;
-  top: sizes.$regular;
-  z-index: 100;
-  margin-right: 3px;
 }
